[Platform] Split out PlatformShim; Add a small readme
